+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidyr (development version)
 
+* `fill()` gains a `.by` argument as an alternative to `dplyr::group_by()` for
+  applying the fill per group, similar to `nest(.by =)` and
+  `dplyr::mutate(.by =)` (@olivroy, #1439).
+
 * `unchop()` produces a more helpful error message when columns cannot be cast
   to `ptype` (@mgirlich, #1477).